Review Request 4576: fix pls deployment test to clean up model summary

+++ le-pls/src/test/java/com/latticeengines/pls/controller/ModelCopyResourceDeploymentTestNG.java
@@ -89,7 +89,9 @@ public class ModelCopyResourceDeploymentTestNG extends PlsDeploymentTestNGBase {
         HdfsUtils.rmdir(yarnConfiguration, customerBase + tenant1.getId());
         HdfsUtils.rmdir(yarnConfiguration, customerBase + tenant2.getId());
         setupSecurityContext(tenant1);
-        modelSummaryEntityMgr.deleteByModelId("ms__20a331e9-f18b-4358-8023-e44a36cb17d1-testWork");
+        if (modelSummaryEntityMgr.getByModelId("ms__20a331e9-f18b-4358-8023-e44a36cb17d1-testWork") != null) {
+            modelSummaryEntityMgr.deleteByModelId("ms__20a331e9-f18b-4358-8023-e44a36cb17d1-testWork");
+        }
     }
 
     private void setupTwoTenants() throws KeyManagementException, NoSuchAlgorithmException {
